+++ fastly/ngwaf_rule_helpers.go
@@ -1,7 +1,7 @@
 package fastly
 
 import (
-	"github.com/fastly/go-fastly/v11/fastly/ngwaf/v1/rules"
+	"github.com/fastly/go-fastly/v12/fastly/ngwaf/v1/rules"
 )
 
 func expandNGWAFRuleConditionsGeneric(