+++ fastly/resource_fastly_ngwaf_alert_datadog_integration.go
@@ -10,8 +10,8 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 
-	gofastly "github.com/fastly/go-fastly/v11/fastly"
-	ddalerts "github.com/fastly/go-fastly/v11/fastly/ngwaf/v1/workspaces/alerts/datadog"
+	gofastly "github.com/fastly/go-fastly/v12/fastly"
+	ddalerts "github.com/fastly/go-fastly/v12/fastly/ngwaf/v1/workspaces/alerts/datadog"
 )
 
 func resourceFastlyNGWAFAlertDatadogIntegration() *schema.Resource {
@@ -34,7 +34,7 @@ func resourceFastlyNGWAFAlertDatadogIntegration() *schema.Resource {
 				Required:     true,
 				Type:         schema.TypeString,
 				ValidateFunc: validation.StringLenBetween(9, 9),
-				Sensitive:    true,
+				Sensitive:    !DisplaySensitiveFields,
 			},
 			"site": {
 				Description:  "The Datadog site.",