+++ fastly/resource_fastly_ngwaf_alert_slack_integration_test.go
@@ -10,7 +10,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 
-	webhookAlerts "github.com/fastly/go-fastly/v11/fastly/ngwaf/v1/workspaces/alerts/slack"
+	webhookAlerts "github.com/fastly/go-fastly/v12/fastly/ngwaf/v1/workspaces/alerts/slack"
 )
 
 const (